move user details to components folder
